requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
allure-pytest==2.14.0
allure-python-commons==2.14.0
attrs==23.2.0
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
fake-useragent==1.5.1
Faker==37.3.0
Flask==3.1.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mock-open==1.4.0
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-sugar==1.0.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
termcolor==3.0.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.9.4
